pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
gunicorn = "^20.1.0"
requests = "^2.28.1"
Flask = "^2.1.2"
pytest = "^7.1.2"
pymongo = "^4.3.2"
mongomock = "^4.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
